+++ src/ios/index.js
@@ -1,0 +1,5 @@
+const { iosRuntimeList } = require("./functions");
+
+module.exports = {
+  iosRuntimeList,
+};
